design changes
--header is now global
--disabled install buttons if verisurf directory not found
