+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "verisurf-nightly",
-  "version": "1.0.21",
+  "version": "1.0.23",
   "main": "public/Main.js",
   "repository": {
     "type": "git",
@@ -20,9 +20,11 @@
     "react": "^16.12.0",
     "react-apexcharts": "^1.3.6",
     "react-dom": "^16.12.0",
+    "react-redux": "^7.1.3",
     "react-reveal": "^1.2.2",
     "react-router-dom": "^5.1.2",
     "react-scripts": "3.3.0",
+    "redux": "^4.0.5",
     "regedit": "^3.0.3",
     "typeface-roboto": "0.0.75"
   },
@@ -74,7 +76,7 @@
       "build/",
       "node_modules/**/*"
     ],
-    "win":{
+    "win": {
       "requestedExecutionLevel": "requireAdministrator"
     },
     "directories": {
@@ -89,5 +91,5 @@
         ]
       }
     ]
-  } 
+  }
 }
