+++ src/pages/Settings.css
@@ -3,6 +3,7 @@
     height: 100vh;
     width: 100%;
     flex-direction: column;
+    margin-top: 5rem;
 }
 
 .settingsHeader{
