+++ src/pages/TestPage.css
@@ -3,7 +3,7 @@
     display: grid;
     min-height: 100%;
     position: relative;
-    margin-top: 2rem;
+    margin-top: 2.5rem;
 }
 
 #currentDownload {
